requirements.txt
streamlit>=1.30.0
snowflake-connector-python>=3.6.0
cryptography>=41.0.0
pandas>=2.0.0
plotly>=5.18.0
anthropic>=0.30.0
